+++ desk/deskfpd.c
@@ -88,7 +88,7 @@ FNODE *fpd_ofind(FNODE *pf, WORD obj)
 static void fl_free(PNODE *pn)
 {
     if (pn->p_fbase)
-        dos_free((LONG)pn->p_fbase);
+        dos_free(pn->p_fbase);
 
     pn->p_fbase = pn->p_flist = NULL;
     pn->p_count = 0;
@@ -283,7 +283,7 @@ FNODE *pn_sort(PNODE *pn)
     }
     pf->f_next = (FNODE *) NULL;
 
-    dos_free((LONG)ml_pfndx);
+    dos_free(ml_pfndx);
 
     return newlist;
 }
@@ -328,7 +328,7 @@ WORD pn_active(PNODE *pn)
 
     if (count == 0)
     {
-        dos_free((LONG)pn->p_fbase);
+        dos_free(pn->p_fbase);
         pn->p_fbase = NULL;
         return 0;
     }
